fix(searchBox): allow searchBox to reuse an `<input>`
This no longer uses React for the `input` display (it still uses it
for the optionnal `PoweredBy` widget).

All HTML attributes already set on the `input` are kept, but sensible
defaults (like `autocorrect`, etc) are added. It listens to the
`keyup` event only to start the search.

There is no `render` method because it does not need to be updated on
results, only triggering it.