feat(templatesConfig): helpers and options transferred to Template
New parameter passed to the `init` and `render` method: `templatesConfig`.
It replaces `templateHelpers`.

This parameter embeds both the `helpers` and the `Hogan` options.
It uses the same chain that the `templateHelpers` previously used to
be passed from a widget to the `Template` component.

The `Template` component now also accepts a `templateDefault` parameter
which allows us to ignore `Hogan`'s `options` if there is no
`template` provided

Closes #99.