feat(searchbox): Make the searchBox BEMish
I've also updated the `poweredBy` as well. I've changed the logic of
adding inline `style` to using a specific class (defined in the
default `css` file). I've switched the display/hide logic. We no
longer need to define a `display=true` in the props to show it, but
a `hidden=true` to hide it instead.

BREAKING CHANGE: The `searchBox` widget now expect
a `cssClasses.{input, poweredBy}`