test(RefinementList): add missing expectJSX
I will PR our travis.yml to run the tests using karma in a real chrome
browser. So that we detect those missings in advance.

The reason this do not fail on regular test:watch task or test task is
that using pure mocha/nodejs env means having a single require context
(thus expectJSX is always present once required once).

But in karma, every test file is a new context.