chore(lib): move main export to lib/
It's a pain to try to unit test root files with karma AND mocha
binary.

Specifying multiples files in subdirectories + ignoring node_modules +
having root test files is really hard to handle.

Karma was not running the index-test.js for instance.

For consistency, this is also better, because now we are in __tests__