chore(warnings): Remove test warning about `type should not be null`
The warning `Warning: React.createElement: type should not be null,
undefined, boolean, or number. It should be a string (for DOM
elements) or a ReactClass (for composite components).` was caused by
calling the test decorator on `<div />` instead of `'div'`.

As the message suggest, we should use strings for basic DOM elements.