docs(dev): Expose all internal webservers to the same ip
Jekyll now is exposed on 0.0.0.0, which means that you can connect to
it through your machine ip. Webpack-dev-server was already doing it,
as well as the guard livereload.

All the instantsearch css, js and livereload are now loaded through
ugly `document.write` calls in dev mode, to be able to know which is
the current ip that is used to access the website.

This works both when testing locally on desktop, on mobile through the
same wifi network, and through an ngrok tunnel.