fix(url-sync): adds indexName in the helper configuration
This solves a bug found using safari and that may or may not
break an implementation based on InstantSearch.

Safari triggers a pop state event on load, which triggers a new
search. But at this point the index is not in the initial
configuration, so this search will fail with a 400, indexName not
valid.

With this fix the value for the index name is taken from the parameters
and its value is added to the initial search configuration.