feat(priceRanges): add currency option
This is a feature and also a fix.

Previously, we had labels.currency but it was not setting the currency
in the item template.

So you had to provide your own template item, and even doing so could
not access any `currency` template data.

No we only provide opts.currency (no more labels.currency, but we are
backward compatible).

And `currency` is a template data.