fix(url-sync): always decode incoming query string
Issue: we decided not to put encoded URIS inside the url bar because
it looks very bad (%23.. everywhere) since we are using things like
"&nR[popularity][>=][0]=0".

But then, when copy pasting a URL with "&nR[popularity][>=][0]=0"
things gets messy: GUI clients

As for copy pasting the url from the urlbar or right click a link and
copy from the context menu then paste it anywhere => it's the
responsibility of the pasted-to client to do things well (escape, make
it clickable).

But then, on our side, we can be sent encoded URIs, so we need to
decode them first before passing them to the helper.

This should be harmless. What do you think?

fixes #848