fix(Template): stop leaking `data="[object Object]"` attributes in production builds
Avoid `<div data="[object Object]">` being inserted in the DOM in
production builds.

The bug was introduced by
https://github.com/algolia/instantsearch.js/commit/fcfb990c9d059f59389b753b5abdecfba0dcd0f8
where we started using transform-react-remove-prop-types.

We cannot rely on PropTypes in production env because they are
removed.

We ask the dev to be explicit in what he's doing: use rootProps versus
forwarding everything a bit magically.

fixes #899