fix(ci): Testing Travis
Hotfix:

The script to validate PRs was really confusing about what being a PR
meant or not. This version should work better.

Basically it will succeed when not on Travis, reject PRs not on
develop, except if they are on master with either a `docs()` or
a `hotfix` mention.

So this very own PR, even if done on master, should just be valid.