fix(searchBox): do not update the input when focused
When we have an external update:
- search.helper.setQuery
- debounce option used on searchBox (resulting in an "external"
update: search happens after some time)

Then if the input is focused (user typing), we should not try to
update it. It will always result in a bad behavior.

Thus, it means we do not handle use cases like "When the user is
focused, I want to programmatically update is query". If you want to
do that, you would have to blur the input, update, focus.

But that's an edge case and bad UX I believe.

fixes #944