Add titles to widgets' images on examples page
Unless you are familiar with the icons of each widget, the current images are hard to understand. I added the title so you at least get a tooltip - making it easier to find the widget in the doc afterwards.
Also, I'm not really sure about the syntax and didn't want to deploy everything to test, so tell me if I should make some changes.