fix(rangeSlider): step accepts a float value
Before this commit we were not able to use float values for step, to
be able to have a slider with float values.

Rounding was introduced when fixing
https://github.com/algolia/instantsearch.js/pull/477.

Now that step = 1 by default we do not need rounding automatically. If
the user is careful about all the other widgets using the same
attribute to only refine rounded values then he's fine.