fix(urlSync): only start watching for changes at first render
Before this commit, if one of the widget was making helper changes in
his init(), we would trigger a change event that would update the url.

What we want is that everything done before the first render should be
considered as start configuration and not synced into the url UNLESS
there are new differences (user changes the query).

It also fixed a nasty IE10/11 bug where the input event would be
triggered as soon as a placeholder is set into the DOM or
programmatically.