docs(transformData): Only document explicit objects for `transformData`
We can currently pass for `transformData` either an object or
a function. If passed an function, it will be applied on every
template, while if passing an object, each key function will only be
applied to the template with a matching name.

We think the "top level" call to a function creates more issues than
an explicit definition. It usually creates weird effects when the
method is applied to a "no result" set, resulting in errors. To avoid
that, we will stop documenting the feature, and remove it completely
in a next major version.

I've updated the documentation to reflect this change, ie always
documenting `transformData.item` instead of just `transformData` as
the place to put the custom method. I've updated the `usage` errors
accordingly as well.

I also fixed a few errors in the docs, like removing empty default
values (it was only adding noise to the docs, and was not consistent).

I've also fixed an inconsistency between the documentation and the
code in the `currentRefinedValues` widget. It was previously not
possible to pass a `transformData.item` (this resulted in an error),
while passing a `transformData` directly was accepted, but didn't
produce any effect. We can now correctly pass `transformData.item`.