fix(numericRefinementList): Correctly apply active class
Fixes #1010

The `shouldComponentUpdate` of `RefinementList` was checking if the
`facetValues` from the props where actually changed before issuing
a new rendering. It turns out that in `numericRefinementList`, the
new `facetValues` was always equal to the already set one. More
precisely, the filters seemed to already be set in the props before
reaching `shouldComponentUpdate`.

Turns out that we were editing a variable in place instead of working
on a copy so the internal props where changing without the component
really noticing.

I updated the `render` method of the `numericRefinementList` to pass
a new array instead of working on the same copy over and over again.

I refactored the `shouldComponentUpdate` and `_toggleRefinement`
methods by splitting in several variables in my quest to find the
issue. Thought I might keep it that way.

I added an explicit regression test to make sure we are really not
touching existing variables.

Now the current and next props are really different, and the full
render is propagated and the correct classes applied.# Possible types : feat, fix, refactor, chore, style, perf, test, docs