docs(faq): Add "hide results on init"
I suggest adding a FAQ section to the documentation where we could put
some of the most common use cases of instantsearch that are either not
yet into the core, or are not suitable to be put into the core.

I just had a user asking me how to make instantsearch not display the
first results, and only display them on the first keystroke.

I wrote a small explanation, with code example and added it to the
docs.

Let me know what you think