docs(faq): Add explanation on default filters
Following #1026 I also added an entry on how to set default filters on
the first loading of instantsearch. This question often pop ups in
support and we were never really explaining how to do it.

I tried to keep it as simple as possible for most cases when you
already have a widget in place on that specific attribute, and only
need to set default values.

I also explained how to do it on attributes on which you do not have
a widget, with a bit more config.

Let me know what you think