docs(header): Fix the header when scrolling
I've had feedbacks that when discovering instantsearch.js for the
first time, people scroll down on the home page and once they are at
the bottom they don't know where to go because the menu disappeared.
They thought it was just a promo page, with no more content, so they
closed it.

I've set the header as `position: fixed` like on the documentation
page. Seems like a recent commit did remove the fixed header by
mistake.