fix(lite): use lite algoliasearch build (js client)
fixes #1024 by forcing to load /lite and avoid bad resolve of the
browser field.

This could break if people were using search.client to do things like
indexing. But we do document this as an API entry. We only have an
infinite scroll tutorial using the client passed to the helper. Thus I
had to make browse and browseFrom part of the lite build.