fix(poweredBy): Let users define their own poweredBy template
Fixes: #1047

You can now pass an object to `poweredBy` like this:

```javascript
poweredBy: {
  template: 'Link to {{data.url}}', // Also accepts function
  cssClasses: {
    root: 'applied to wrapper',
    link: 'applied to link'
  }
}
```

Everything is handled directly in the widget, without the need for
a React component. This was overkill because we never update the
render of the `poweredBy` so it was easier to just generate the
template at init time without going through all the process of
creating a template.

I've also added more thorough tests. Users will be able to
independently change CSS classes or template.