feat(header): Pass count of current refined filters in header
Partially fixes #1013

The goal of the PR was to make the `count` variable accessible to the
header template of the `refinementList` widget. This `count` is the
number of active filters applied to the facet.

To do so the first step was to make the `header` template dynamic.
Before the PR, it was static, meaning that it was defined at the
widget init and never updated. I changed that part so the header
template is re-rendered on every render of its component.

It also meant that I had to remove the custom `shouldComponentUpdate`
method. The method was checking if the component was collapsed or not
before triggering the render. As we now always need to update the
render of the header, being collapsed or not, I removed the check.
I think it might even fix #1041 as well.

To pass template data to the header, you now need to pass the
`headerFooterData` prop to the component that is wrapped by
`headerFooterHOC`. It's an object with a key `header` containing the
data you want to pass to each. It could also potentially have a key
`footer`, but the footer part is not implemented yet because we had no
need for it.

Improvements and discussions:

- The `footer` template is still static. It would be trivial to set it
  as dynamic, but I left it like this until we have a use-case.
- Only the `refinementList` widget takes advantage of this feature. It
  would not be too hard to pass dynamic data to headers in other
  widgets, but I left that to other PRs.
- I will tackle the other issue of adding a CSS class on the root on
  another PR.