feat(hits): Add a `__position` attribute to data passed to items
Fixes #903

The change itself was easy to do, but it broke all the tests because
we were testing a full JSX output and we now had one more key passed
to the `data`.

It took this as an opportunity to refactor the Hits tests uzing enzyme
to make them easier to read. It allow me to spot one or two
inconsistencies in the way the `root` class was added or not depending
on the kind of template used.