API docs, website bootstrap, main expose bootstrap
This is all very WIP but should be a good base to be able to move
forward with implementing more features.

I tried to create most of the needed widgets we need to implement,
without detailing them most of the time.

To detail them, have a look at the current instantsearch.js V1 and
always take a look at the widget:nameOfWidget label on github so that
you better understand what we need to change.

I created some widgets which are direct ripoff searchkit ones because
their naming of ranges is the right one. A range might be by default a
slider, but not always (think: createRange).

Our next goals should be:
- implement those widgets with the best API you can think of while being
  inspired with what exists and the current issues in IS (label:
widget:..)
- continue documenting all widgets and HOC (have a look at
  createSearchBox and
  SearchBox documentations to see a complete example)
- continue bootstraping the website, have a look a site.js
- HAVE FUN.

I removed the dependencies of the examples/ folder from the main
package.json, please create an examples/webpack.config.js and
examples/package.json so that we do not bloat the main package.json too
much.

To start the website: npm start