fix(bundle): switch back to React by default, create a preact build
Create a specific preact build instead of defaulting to preact.

We introduced a breaking change for some users when moving to preact
by default (in #1224).

Since we allow React components as templates and some users might be
doing this on the frontend (without preact), let's create a preact
build for now while we investigate.