fix(rangeSlider): round pips numbers when step is integer
The default rounding strategy for pips in nouislider is to use
Math.round().

We used Number(v).toLocaleString() but on safari this led to weird
results because 10.100001.toLocaleString() => '10.1000001' and on
chrome it's  '10' for example.

Thus now we will Math.round the number before passing it to
toLocaleString only if the step parameter of the slider is an integer.

We use an npm module for Number.isInteger() because no IE nor safari
version has it.

fixes #1254