fix(pagination): reset on other filter + no toggle
The bug is caused by the use of two conflicting semantics
with the same set of changes. Previously we had "if the value for the
page, reset the page number", which lead to reseting the pager to 0
if we were selecting the same page (the toggle effect). I previously
removed this behaviour, but it introduced the imposibility to reset
the pager to 0 when changing any other filter (setting the value was
the trigger).

In order to fix this issue I introduce a new type of value to
express the intent of using the same value without changing the page.
This way we have two distinctly defined intent and values:
 - same value as numbers -> reset to 0
 - value with isSamePage flag -> change to page value

It relies on `valuesOf`, which is not that nice and introduces the use
of the dreaded `==`. At least, I can say that I found a legit use case
for it. Bottom line, I'm open to suggestion here :)