feat(sortBy): implement default sortBy, remove option (#1549)
Fix #1529. The motivation is that any sort can be applied directly in
the component and does not require an API for that (the default behavior
is good given our understanding of search principle)

BREAKING CHANGES: sortBy option does not exists anymore. Please use a connector or open a new issue with your usecase for it.