chore(dev): tune dev env for 200ms updates (#1584)
Before this commit, editting a single CSS file or markdown guide took
8s on my computer.

Now it takes 200ms.

It's ugly as a fix, it would need to be baked in. Having a good static
site builder AND a very nice dev experience on it is key (but hard) to
ease the contribution/iteration.