docs(website): improve the readability of required (or important) options
Often times, I feel a bit lost while reading the documentation because
all the options' descriptions re collapsed, resulting in me considering
all the options with the same importance.

I feel like this is definitely something we want for templates &
cssClasses but some other options (even optional ones) are used a lot
and should be emphasized.