docs(review): complete docs review
- /guide/ and /examples/ instead of /guide.html /examples.html so that
if you are on /guide/Installation.html and try to go on /guide/, it
works
- moved advanced topics
- reorganized the guide so it follows a progressive learning path
- read the guide entirely and made adjustement as if I was a new
user/reader
- made small adjustements on the rendering
- removed the need for @kind + @category, only kept @kind
(InstantSearch component is OK as a widget, it's not really a widget
but 90% of our components are widgets so it's easier to only talk
about widgets for users)
- removed headings BUT from the getting started, pages are too small
to have headings (and sometime repetitive headings). Did not feel the
pain not having them, reduces noise

I have other ideas how to make the docs better, will bring them to
discuss once we are live. Right now we should be safe for a first
release content wise.

fixes #1599