fix(Highlight): remove infinite loop
Before this commit, when using `<Highlight>` inside `<Hit>` or any
`connectHits` we ended up in an infinite loop.

This loop was hard to notice because it was doing some network at each
iteration.

We will open new issues related to this bug because it's not just the
`Highlight` widget fault, it's the current design of computing
searchParameters and dynamic mounting that is failing.

This needs a deeper investigation and is actually why server side
today is not really supported.