fix(range): when unfinite numbers are passed, use search results instead.
We discovered this behavior when using the rheostat slider widget. For an unknown reason when you have the same value for Min/Max, the slider gaves you NaN/NaN instead of the real value.
The consequence for react-instantsearch is that when the 'shouldComponentUpdate' hook of the range connector is triggered, equality will never be reached as the currentRefinement contains 'null' value. see #1727