fix(createInstantSearchManager): drop outdated response (#1765)
The helper is implemented in a way that it will drop outdated requests.
This change makes react-instantsearch use the event model of the helper
that let it implement this mechanism that ensures reliability in the
algolia answers. It also results in some fewer instanciation of objects
for callbacks and the base search parameters.