feat(pending-search): let the dev know the state of the search requests queue
New feature *pending-search*: let the dev know the pending state

With this feature, the developper will be able to know what is the state
of the current searches. If there is a pending search, the dev can
decide to display a visual feedback so that the user something is
going on. This is intended to be able to let the user know that there is
an on-going search, and prevent their frustration.

It implements two things:
 - a new method called `hasPendingRequests` that returns `true` if there is
a search pending.
 - a new event `searchQueueEmpty` that is fired when all the requests have
been served or cancelled