fix(autoHideContainer): dont prevent render with `shouldComponentUpdate`
If you click on a DOM element which has a state (ex radio button),
React wont see this state change because children component of `autoHideContainer`
wont be updated because of the implementation of `shouldComponentUpdate`.

It's a weird bug where UI and state aren't in sync because the navigator
saves on it's own the state of the DOM element.

fixes https://github.com/algolia/instantsearch.js/issues/2066