fix(show-more): should hide button when show more is not available
As noticed by @timkelty in #2160, showMore button should not be rendered
if there are not enough items to display based on the limits. With the
new connector API, this means that we should add a new parameter to
the renderer to give the information that this button should be
displayed.

fixes #2160