feat(searchFunction): make search function provide a better API
* feat(searchFunction): Offer full helper support in search function

This implementation offers:
 - a full helper to modify the parameters before sending the request
 - all the other search function (searchOnce, searchForFacetValues)
 - compatibility with async code (no oops effect)

* fix(slider): Make sure the slider doesn't update the state twice

This commit adds constraints to make sure that a refine call doesn't
trigger a new state update and search. It also adds undefined as a
value that can be passed as "no boundary selected" for min and max
in the refine function (slider connector).

* fix(searchFunction): rm unnecessary state update in internals

* chore(naming): change name of search helper function

* chore(test): test searchFunction in a less altered env

The previous setup was not compatible with the kind of manipulation done
for implementing the `searchFunction`.