fix(selector): root classname is applied twice
The implementation of the React `selector` was applying the root css
classname that is already applied by the headerfooter. In order to
prevent regressions and let user customize the <select>, this patch adds
a new key to customize cssClasses in Selector bases widgets, and this
new key has the root value to avoid breaking the current implementations.

Fix #2396 #2397