fix(selector): root classname is applied twice (#2423)
* fix(selector): root classname is applied twice
The implementation of the React `selector` was applying the root css
classname that is already applied by the headerfooter. In order to
prevent regressions and let user customize the <select>, this patch adds
a new key to customize cssClasses in Selector bases widgets, and this
new key has the root value to avoid breaking the current implementations.

Fix #2396 #2397

* chore(test): update snapshot
* chore(functional-tests): add form-control using select
The new cssClasses key so that we check that all 3 can have custom css
classes. Interestingly the key was used but not implemented already :/

* chore(test): update test to validate the usage of the new key