feat(core): the core provides information about stalled search
This commit adds a new step that can update the UI. In this change,
the widgets are rendered when there are new results (previous behaviour)
and when the search is found to be stalled. The information is
propagated through a new option `searchMetadata` to the render method of
the widgets. This is an object that contains the flag `isStalledSearch`
that is set to `true` when there are still pending search after a delay
(for now 200ms).