chore(linting): target correct folder (#2566)
* chore(linting): target correct folder
ignore correct things now
make hex rule a warning for now because we aren't consistent with it, we can fix it later when we want less warnings

closes #2445

* chore(sass-lint): update rules based on our usage

* chore(scss): update files based on current rules

There are still two kinds of lint errors that I'm not sure about. I'm
not really in favor of forcing the nesting all the time. And not using
qualifying elements can have unforeseen issues.

* chore(theme): fix sass-lint errors

* chore(style): enable empty-line-between-blocks