feat(core): provide information about stalled search to widgets (#2569)
* feat(core): the core provides information about stalled search
This commit adds a new step that can update the UI. In this change,
the widgets are rendered when there are new results (previous behaviour)
and when the search is found to be stalled. The information is
propagated through a new option `searchMetadata` to the render method of
the widgets. This is an object that contains the flag `isStalledSearch`
that is set to `true` when there are still pending search after a delay
(for now 200ms).

* feat(searchbox): implements a loading indicator
* chore(lint): fix prettier and eslint check
* feat(core): add option to configure stalled search delay
* feat(searchbox): add options to set a loading indicator
* chore: update test to match the lib changes
* chore(search-box-test): remove sinon
* chore(connectSearchBox): remove sinon from test
* chore(doc): add isStalledSearch on connectSearchBox rendering options
* chore(validation): merge commit length should not be checked

Merge commits are automatically generated and are already standardized
by git, therefore there is no point in checking their length. Worst
truncating them would make use loose information.

* test(core): test new render step with stalled search
* chore(searchbox): test with input tag container
* chore(doc): fix inconsistencies and add details
* chore(searchbox): use `.ais-stalled-search` and fix use of input
* chore(doc): fix content
* chore: merge with 2.3, update snapshot