chore(package.json): add files entry (algolia/algoliasearch-helper-js#557)
* chore(package.json): add files entry
Right now all files in the directory are published to npm on publish. On the [Yarn site](https://yarn.pm/algoliasearch-helper) you can check which files are included now, which includes the documentation and other files. 

Also removes the .npmignore file, since a whitelist is better than a blacklist
* chore: Delete .npmignore
* chore: remove default files
readme, changelog etc are pushed by default and don't need to be listed
* chore: add index.js