chore(security): remove public tokens
Following security reports, removing public tokens for saucelabs and
argos.

For Saucelabs we are now using https://docs.travis-ci.com/user/jwt which
is the recommended way to have external contributors PRs working (read
documentation on travis).

For Argos jwt is not implemented so it will just be disabled, we need to
be careful while reviewing for now (same setup as react-instantsearch).