chore(pkg): use jsDelivr default file feature
We can reduce the urls we give to users for the JavaScript package by
informing jsDelivr what's the right default file to serve when nothing
provided after the @version.

This is also handy as it does not ties the user and us to a particular
schema to follow over time, if we ever decide to move our build, change
it and rename things, it would be transparent for the user.

For v2 that's "too late" but still good to use it for documentation and
ease of use.

For the theme and style ideally it would be in another package too
("IS.css?" :D)