fix(perf): only compute snappoints when step is provided (#2699)
* chore: do not uglify in dev + better banner
* fix(perf): only compute snappoints when step is provided

This keeps the snapping as it snaps points to the rounded number if
no snappoints are provided. Until rheostat provides a lazy way to fix
the snappoints, there will also be a risk of having an "out of memory"
error because the range of values is too big. There is an open issue
opened for adding this new API: https://github.com/airbnb/rheostat/issues/14

Fix #2662

* chore: fix lint
* chore: fix lint