fix(Template): harden Symbol checks
If a `Symbol` polyfill is present, the instantsearch script presumes
that `Symbol` is fully implemented. However, `Symbol` polyfills cannot
replicate the behavior fully.

This causes the instantsearch script to error out and cause problems on
the page if a `Symbol` polyfill is used.

This commit hardens the `Symbol` checks to make sure real `Symbol`s are
supported, not just a polyfill.

NOTE: The google maps API library is a common library that implements a
`Symbol` polyfill.

@see:
* https://stackoverflow.com/questions/48532850/is-google-maps-js-library-polyfilling-symbol
* https://stackoverflow.com/questions/39803281/get-typeof-of-the-value-in-es6#39805240