fix(connectSortBySelector): correct behavior of default index
Previously the index name used as the default was the one at the init.
However this one is the one from the URL if provided. We should rely on
the one provided on the instance of InstantSearch.

For this purpose, I had to provide also an instance of InstantSearch in
the test.