fix(clear-all): apply excludeAttribute correctly
This fix actually contains a rewrite of the underlying functions
that will compute the list of attributes to clear. The complete required
behaviour was spread both in the connectors and the utility functions.
In order to simplify the code and make it easier to test, this
refactoring was needed.

A next step for this is to use a single connnector for current refined
values and clear all widgets.