fix(infinite-hits): fix #2543
The infinite hits widget is based internally on the hits widget which
accepts the allItems template. However this template is not usable for
the purpose of the infinite hits widget.

In order to prevent users from using this buggy and yet accepted option
this PR generates a exception if we detect this particular template.

Fix #2543