fix(release): provide interactive TTY for npm publish (#3053)
When we enforce 2-factor auth on our accounts, NPM will asks for the OTP
which requires an interactive terminal. This is only achievable using
child_process as per the shellJS FAQ: https://github.com/shelljs/shelljs/wiki/FAQ#running-interactive-programs-with-exec

This also reorders the commands so that we can do the rollback without
making the code more complex that it is already.