feat(InstantSearch): call addAlgoliaAgent if possible (algolia/vue-instantsearch#537)
* feat(InstantSearch): call addAlgoliAgent if possible

still to do: read it from the package.json etc. I have to figure out how it definitely doesn't go fully in the source code. I have a babel macro written for it, but we use buble and not babel here

* feat(version): get from package.json

rollup only conditionally will pick keys from package.json.

If someone imports the files themselves this might cause a problem, but isn't exactly supported for now so that's fine :)

* test: no `v` in version

* chore: move logic inline