feat(sffv): throw an error if it's called and the client doesn't have the functions (algolia/algoliasearch-helper-js#623)
* feat(sffv): throw an error if it's called and the client doesn't have the functions

We will throw an error if `client.searchForFacetValues` is not a function and `client.initIndex` neither. We won't explicitly test if `client.initIndex(...).searchForFacetValues` is there, but that's the old signature anyway, so it doesn't explicitly need this clearer error. We also don't assert that they are returning Promises, but i haven't seen that problem yet, so i don't think it's necessary to throw a specific error there.

* refactor: feedback

* chore: make test less focused

* test: assert on the test message