fix(hits): transform items after escaping
This applies the items transformation after escaping.
It fixes issues when we transform highlighted values.

An example could be that the highlighted value gets uppercased, and therefore change the default tag "__ais-highlight__" to "__AIS-HIGHLIGHT__" which doesn't get caught by the regex to be replaced by a `mark` tag.

The fix is implemented for both Hits and InfiniteHits connectors.

Closes #3250