fix(toggleRefinement): provide `name` to templates (#3303)
## Context

> The default template uses the `name` attribute but it’s not provided to the template. It renders something blank on the page. We should find a “good“ default value or we make the `template.labelText` required. Previously it was to the label that was a required option.
> – @samouss 

## This solution

We pass the attribute (called `name`) to the value passed by the connector `connectToggleRefinement`. The refinement now has a `name` property usable in the template.

If your attribute is called `free_shipping`, the default template will display "free_shipping". You will then need to specify `templates.labelText` to change it to "Free shipping".