feat(tourism): migrate example (#3373)
Move the example to `instantsearch.js@3.0.0`. Note that depending the size of the page the map can zoom out. It's a know issue that we have with Google Maps and float dimensions (see [#1737](https://github.com/algolia/react-instantsearch/issues/1737)). It's possible to fix it with fixed (integer) dimension. We gonna try to fix it directly inside the lib but we didn't have the time to take a look yet.