feat(connectCurrentRefinements): add a root label (PART 1)
This will always be identical to the attribute, and will be the thing which will be displayed.

The reason why you'd want a key like this is if you e.g. want to rename a single key or transform all keys to be upper case, without actually changing which attribute will be changed.