fix(tools): use commonjs in bump-package-version.js
It seems before 92966391b1b2d2870827f164f1046d9a0c544efb the script was
run in developement env with no modules setting, which mean it defaulted
to the poorly documented `auto`.
Looking at their codebase it seems `auto` is simply `commonjs`.  This
explains why it worked before.

We could have used BABEL_ENV=cjs in front of the babel-node command but
I've moved this file to cjs as it's more simple and is consistent with
other release scripts.