chore(jest): put config top-level
Some tools (like the jest plugin for vscode) otherwise would need to be configured for this path too, which is a bit of a hassle.

The other option is of course to set those settings in the workspace, but it's easy to forget since we don't all use the tool.