fix(merge): change implementation  (algolia/algoliasearch-helper-js#716)
* fix(merge): use .toStrictEqual instead of .toEqual in tests

cf. https://github.com/facebook/jest/issues/711

`undefined` values being critical to in assessing merge behaviour this
is necessary

* fix(merge): specify description

add meaning of array-like

* fix(merge): change implementation

- split the implem into merge(target, ...sources) and _merge(target, source), I think it's more simple this way.
- I reorganized some of the conditions

The description and the test suit should not change. Implementation has
changed but behaviour and limitation is the same (Note that the lodash
inspired test suite for merge should be passing).

* fix: extra-space

Co-Authored-By: Yannick Croissant <yannick.croissant@gmail.com>

* fix: rename isObjectOrArray to isObjectOrArrayOrFunction