feat(index): reset page of child indexes
Description

This adds support for resetting pagination for child indexes whenever the state changes in a parent (except when only the `page` search parameter changes). This information is given by the helper `change` event: `isPageReset`.

This is done recursively so that we reset the pagination for _all_ child index widgets. Note that this is a noop when the pagination is not set in the children (because we base the condition on `isPageReset`).

Example

```
GIVEN

index1: page 1
  index2: page 3
    index3: page 4
  index4: page 2

WHEN

query change on index2

THEN

index1: page 1
  index2: page 0
    index3: page 0
  index4: page 2
```